% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsab_palette.R
\name{ds_palette}
\alias{ds_palette}
\title{Get DSAB brand colours}
\usage{
ds_palette(type = c("solid", "light"), as_named = TRUE)
}
\arguments{
\item{type}{\code{character} \code{"solid"} (default) or \code{"light"}.}

\item{as_named}{\code{logical} If \code{FALSE}, return an unnamed vector
(useful for discrete scales). Default is \code{TRUE}.}
}
\value{
A character vector of hex colour codes.
}
\description{
Returns a named character vector of DSAB brand colours.
}
