% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolinska_palette.R
\name{ki_palettes}
\alias{ki_palettes}
\title{Karolinska Institutets official color palettes}
\usage{
ki_palettes(
  palettename = "ki_primary",
  n,
  type = c("discrete", "continuous"),
  order = c("gradient", "hue", "color", "original")
)
}
\arguments{
\item{palettename}{The palette used. Default is \code{"ki_primary"}. Options are \code{c("ki_primary", "ki_function_1", "ki_function_2", "ki_function_3")}.}

\item{n}{Number of colors to return. Defaults to the full palette length.}

\item{type}{\code{"discrete"} or \code{"continuous"}.}

\item{order}{The ordering of colors. Default is \code{"gradient"}. Options are \code{c("gradient", "hue", "color", "original")}.}
}
\description{
The colour palettes can be found on KI's webpages https://medarbetare.ki.se/farger-i-kis-grafiska-profil
}
