% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolinska_palette.R
\name{scale_colour_ki_d}
\alias{scale_colour_ki_d}
\alias{scale_color_ki_d}
\title{KI discrete colour scale}
\usage{
scale_colour_ki_d(palettename = "ki_primary", order = "gradient", ...)

scale_color_ki_d(palettename = "ki_primary", order = "gradient", ...)
}
\arguments{
\item{palettename}{KI palette name. Default \code{"ki_primary"}.}

\item{order}{Colour ordering. Default \code{"gradient"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{scale_colour_manual}()}.}
}
\description{
KI discrete colour scale
}
