% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsab_palette.R
\name{scale_fill_ds}
\alias{scale_fill_ds}
\title{DSAB brand fill scale}
\usage{
scale_fill_ds(type = c("solid", "light"), ...)
}
\arguments{
\item{type}{\code{character} \code{"solid"} (default) or \code{"light"}.}

\item{...}{Additional arguments passed to
\code{\link[ggplot2]{discrete_scale}()}.}
}
\value{
A ggplot2 scale object.
}
\description{
Discrete fill scale using DSAB brand colours.
}
