% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karolinska_palette.R
\name{scale_fill_ki_c}
\alias{scale_fill_ki_c}
\title{KI continuous fill scale}
\usage{
scale_fill_ki_c(palettename = "ki_primary", order = "gradient", ...)
}
\arguments{
\item{palettename}{KI palette name. Default \code{"ki_primary"}.}

\item{order}{Colour ordering. Default \code{"gradient"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{scale_fill_gradientn}()}.}
}
\description{
KI continuous fill scale
}
