% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsab_palette.R
\name{show_ds_palette}
\alias{show_ds_palette}
\title{Preview DSAB brand palette}
\usage{
show_ds_palette(type = c("solid", "light"), labels = TRUE)
}
\arguments{
\item{type}{\code{character} \code{"solid"} (default) or \code{"light"}.}

\item{labels}{\code{logical} Show hex labels on tiles? Default is \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Draws a tile plot showing the colours and hex values for a DSAB palette.
}
