% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crgg_minimal.R
\name{theme_crgg_minimal}
\alias{theme_crgg_minimal}
\title{CR´s ggplot minimal theme}
\usage{
theme_crgg_minimal(
  base_size = 8,
  base_family = "Palatino",
  horizontal = FALSE,
  black = FALSE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{horizontal}{\code{logical} Horizontal axis lines?}

\item{black}{\code{logical} If \code{TRUE}, use black background, else
use white background. Default is \code{FALSE}.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A white, minimal theme with serif Palatino font.
}
\seealso{
Other themes crgg: 
\code{\link{theme_crgg_standard}()},
\code{\link{theme_ki_standard}()}
}
\concept{themes crgg}
