% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_crgg_standard.R
\name{theme_crgg_standard}
\alias{theme_crgg_standard}
\title{CR´s ggplot standard theme}
\usage{
theme_crgg_standard(
  base_size = 8,
  base_family = "Palatino",
  horizontal = FALSE,
  bgcolor = "white"
)
}
\arguments{
\item{base_size}{\code{integer} Adjusts all fonts and spacing etc. Default is \code{8}.}

\item{base_family}{\code{character} A character vector specifying font. Most standard fonts should installed on your system should work. Recommended are:

\itemize{"Palatino"} (Default)
\itemize{"Helvetica"}
\itemize{"Optima"}
\itemize{"CMU Serif"} (Must be installed from web, but looks darn good.)}

\item{horizontal}{\code{logical} Horizontal axis lines (\code{TRUE/FALSE}).}

\item{bgcolor}{\code{character} Default is \code{white}; white bakground. Options are \code{c("white", "grey", "offwhite", "black")}}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A minimal theme with axis lines. Available in white, black, grey and offwhite.
}
\seealso{
Other themes crgg: 
\code{\link{theme_crgg_minimal}()},
\code{\link{theme_ds_standard}()},
\code{\link{theme_ki_standard}()}
}
\concept{themes crgg}
