% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dsab.R
\name{theme_ds_standard}
\alias{theme_ds_standard}
\title{DSAB minimal axis ggplot2 theme}
\usage{
theme_ds_standard(
  base_size = 11,
  base_family = "Arial",
  axis_line_colour = "black",
  axis_line_width = 0.5,
  legend = FALSE
)
}
\arguments{
\item{base_size}{\code{numeric} Base font size. Default is \code{11}.}

\item{base_family}{\code{character} Font family. Default is \code{"Arial"}.}

\item{axis_line_colour}{\code{character} Colour for axis lines. Default is \code{"black"}.}

\item{axis_line_width}{\code{numeric} Line width for axis lines. Default is \code{0.5}.}

\item{legend}{\code{logical} Show legend? Default is \code{FALSE}.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A clean theme that keeps only axis lines, axis text, and axis titles.
Uses Arial by default. Removes panel grid, background, and other
non-essential elements. Conforms to the DSAB graphical profile.
}
\seealso{
Other themes crgg: 
\code{\link{theme_crgg_minimal}()},
\code{\link{theme_crgg_standard}()},
\code{\link{theme_ki_standard}()}
}
\concept{themes crgg}
