% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_karolinska.R
\name{theme_ki_standard}
\alias{theme_ki_standard}
\title{CR´s Karoliska Institutet ggplot theme}
\usage{
theme_ki_standard(
  base_size = 8,
  base_family = "DM Sans",
  horizontal = TRUE,
  bgcolor = "white"
)
}
\arguments{
\item{base_size}{\code{integer} Adjusts all fonts and spacing etc. Default is \code{8}.}

\item{base_family}{\code{character} A character vector specifying font. For this KI theme default is \code{"DM Sans"}, but can be substituted.
Most standard fonts should installed on your system should work. Recommended are:

\itemize{"DM Sans"} (Default)
\itemize{"Adobe Caslon Pro} (Must be separately installed, and may be expensive, but oh so beautiful.)
\itemize{"Helvetica"}
\itemize{"Optima"}
\itemize{"CMU Serif"} (Must be installed from web, but looks darn good.)}

\item{horizontal}{\code{logical} Horizontal axis lines (\code{TRUE/FALSE}).}

\item{bgcolor}{\code{character} Default is \code{white}; white bakground. Options are \code{c("white", "grey", "offwhite", "black")}}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A minimal theme with axis lines. Available in white, black, grey and offwhite. Uses KIs official font DM Sans (Which you should install separately at https://fonts.google.com/specimen/DM+Sans)
}
\seealso{
Other themes crgg: 
\code{\link{theme_crgg_minimal}()},
\code{\link{theme_crgg_standard}()}
}
\concept{themes crgg}
