% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dsab.R
\name{use_ds_theme}
\alias{use_ds_theme}
\title{Set DSAB theme as global default}
\usage{
use_ds_theme(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{theme_ds_standard}()}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Convenience function that calls \code{\link[ggplot2]{theme_set}()} with
\code{\link{theme_ds_standard}()}.
}
