% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dsab.R
\name{theme_ds_standard}
\alias{theme_ds_standard}
\title{CR´s Karoliska Institutet ggplot theme}
\usage{
theme_ds_standard(
  base_size = 8,
  base_family = "Arial",
  horizontal = TRUE,
  bgcolor = "white"
)
}
\arguments{
\item{base_size}{\code{integer} Adjusts all fonts and spacing etc. Default is \code{8}.}

\item{base_family}{\code{character} A character vector specifying font. For this DS theme default is \code{"Arial"}, but can be substituted.
Most standard fonts should installed on your system should work. Recommended are:

\itemize{"Arial"} (Default sans serif font)
\itemize{"Times New Roman"} (Default serif font.)}

\item{horizontal}{\code{logical} Horizontal axis lines (\code{TRUE/FALSE}).}

\item{bgcolor}{\code{character} Default is \code{white}; white bakground. Options are \code{c("white", "grey", "offwhite", "black")}}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A minimal theme with axis lines. Available in white, black, grey and offwhite. Uses KIs official font DM Sans (Which you should install separately at https://fonts.google.com/specimen/DM+Sans)
}
\seealso{
Other themes crgg: 
\code{\link{theme_crgg_minimal}()},
\code{\link{theme_crgg_standard}()},
\code{\link{theme_ki_standard}()}
}
\concept{themes crgg}
